Use attribute_not_exists() condition to avoid overwrite on save.
This replaces the older primary-key-not-null pattern that we
implemented using the deprecated expression syntax. It is now the
recommended way to avoid overwriting an existing item in put - see
https://boto3.readthedocs.io/en/latest/reference/services/dynamodb.html#DynamoDB.Client.put_item

A bonus is that the `moto` library (which is often used as an in-memory
implementation for testing) implements the attribute_not_exists()
semantics, whereas it does not implement is-null tests for a missing
item.